Completed request changes :D
